Move webhook processing into utilities
