+++ apps/web/utils/webhook/google/process-history-item.ts
@@ -1,9 +1,9 @@
 import type { gmail_v1 } from "@googleapis/gmail";
-import type { ProcessHistoryOptions } from "@/app/api/google/webhook/types";
-import { HistoryEventType } from "@/app/api/google/webhook/types";
+import type { ProcessHistoryOptions } from "@/utils/webhook/google/types";
+import { HistoryEventType } from "@/utils/webhook/google/types";
 import { createEmailProvider } from "@/utils/email/provider";
-import { handleLabelRemovedEvent } from "@/app/api/google/webhook/process-label-removed-event";
-import { handleLabelAddedEvent } from "@/app/api/google/webhook/process-label-added-event";
+import { handleLabelRemovedEvent } from "@/utils/webhook/google/process-label-removed-event";
+import { handleLabelAddedEvent } from "@/utils/webhook/google/process-label-added-event";
 import { processHistoryItem as processHistoryItemShared } from "@/utils/webhook/process-history-item";
 import { markMessageAsProcessing } from "@/utils/redis/message-processing";
 import { GmailLabel } from "@/utils/gmail/label";
